Backend: Move the location of some events to more appropriate packages
