--- src/main/java/at/hannibal2/skyhanni/events/DungeonStartEvent.kt
@@ -1,6 +1,0 @@
-package at.hannibal2.skyhanni.events
-
-import at.hannibal2.skyhanni.api.event.SkyHanniEvent
-
-@Suppress("UnusedPrivateProperty")
-class DungeonStartEvent(dungeonFloor: String) : SkyHanniEvent()
