+++ src/main/java/at/hannibal2/skyhanni/events/bazaar/BazaarOpenedProductEvent.kt
@@ -1,6 +1,7 @@
-package at.hannibal2.skyhanni.events
+package at.hannibal2.skyhanni.events.bazaar
 
 import at.hannibal2.skyhanni.api.event.SkyHanniEvent
+import at.hannibal2.skyhanni.events.InventoryFullyOpenedEvent
 import at.hannibal2.skyhanni.utils.NEUInternalName
 
 class BazaarOpenedProductEvent(val openedProduct: NEUInternalName?, val inventoryOpenEvent: InventoryFullyOpenedEvent) : SkyHanniEvent()
