+++ src/main/java/at/hannibal2/skyhanni/events/diana/BurrowDetectEvent.kt
@@ -1,4 +1,4 @@
-package at.hannibal2.skyhanni.events
+package at.hannibal2.skyhanni.events.diana
 
 import at.hannibal2.skyhanni.api.event.SkyHanniEvent
 import at.hannibal2.skyhanni.features.event.diana.BurrowType
