+++ src/main/java/at/hannibal2/skyhanni/events/dungeon/DungeonBossRoomEnterEvent.kt
@@ -1,4 +1,4 @@
-package at.hannibal2.skyhanni.events
+package at.hannibal2.skyhanni.events.dungeon
 
 import at.hannibal2.skyhanni.api.event.SkyHanniEvent
