+++ src/main/java/at/hannibal2/skyhanni/events/dungeon/DungeonStartEvent.kt
@@ -1,0 +1,5 @@
+package at.hannibal2.skyhanni.events.dungeon
+
+import at.hannibal2.skyhanni.api.event.SkyHanniEvent
+
+class DungeonStartEvent(val dungeonFloor: String) : SkyHanniEvent()
