+++ src/main/java/at/hannibal2/skyhanni/mixins/transformers/MixinNEUOverlay.java
@@ -1,6 +1,6 @@
 package at.hannibal2.skyhanni.mixins.transformers;
 
-import at.hannibal2.skyhanni.events.NEURenderEvent;
+import at.hannibal2.skyhanni.api.enoughupdates.NEURenderEvent;
 import io.github.moulberry.notenoughupdates.NEUOverlay;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Pseudo;
